Refactored to base SSV on CSV
CSV and SSV are more closely related formats than TSV (what we had
initially used to base SSV on). Single character delimiter has now
already been replaced in a number of places to solve compiler issues.
Still need to actually write the code to do the parsing of records and
the writing of records correctly with a string delimiter/separator.